Moved the viewer into a separate file
